+++ editor/app/(dev)/canvas/editor.tsx
@@ -1,55 +1,7 @@
 "use client";
 import dynamic from "next/dynamic";
 import React, { useEffect, useState } from "react";
-import "../../desktop.css";
-
-function isElectron(): boolean {
-  // Renderer process
-  if (
-    typeof window !== "undefined" &&
-    typeof window.process === "object" &&
-    // @ts-expect-error electron type
-    window.process.type === "renderer"
-  ) {
-    return true;
-  }
-
-  // Main process
-  if (
-    typeof process !== "undefined" &&
-    typeof process.versions === "object" &&
-    !!process.versions.electron
-  ) {
-    return true;
-  }
-
-  // Detect the user agent when the `nodeIntegration` option is set to false
-  if (
-    typeof navigator === "object" &&
-    typeof navigator.userAgent === "string" &&
-    navigator.userAgent.indexOf("Electron") >= 0
-  ) {
-    return true;
-  }
-
-  return false;
-}
-
-function useIsElectron() {
-  const [_isElectron, setIsElectron] = useState<boolean>(false);
-  useEffect(() => {
-    setIsElectron(isElectron());
-  }, []);
-
-  return _isElectron;
-}
-
-function DesktopDragArea() {
-  const isElectron = useIsElectron();
-
-  if (!isElectron) return null;
-  return <div className="w-full min-h-9 h-9 desktop-drag-area border-b" />;
-}
+import { DesktopDragArea } from "@/components/desktop-drag-area";
 
 const PlaygroundCanvas = dynamic(
   () => import("@/scaffolds/playground-canvas/playground"),
@@ -70,7 +22,7 @@ export default function Editor(
 
   return (
     <div className="w-full h-full flex flex-col overflow-hidden">
-      <DesktopDragArea />
+      <DesktopDragArea className="border-b" />
       <div className="flex-1 overflow-hidden">
         <PlaygroundCanvas {...props} />
       </div>
