+++ editor/app/(workspace)/[org]/[proj]/page.tsx
@@ -14,6 +14,7 @@ import { Skeleton } from "@/components/ui/skeleton";
 import Head from "next/head";
 import { editorlink } from "@/lib/forms/url";
 import { notFound } from "next/navigation";
+import { DesktopDragArea } from "@/components/desktop-drag-area";
 
 export default function FormsDashboardPage({
   params,
@@ -48,59 +49,64 @@ export default function FormsDashboardPage({
         </title>
       </Head>
       <WorkspaceSidebar />
-      <main className="w-full h-full overflow-y-scroll">
-        <div className="container mx-auto">
-          <header className="py-10 flex justify-between">
-            <div>
-              <span className="flex items-center gap-2 text-2xl font-black select-none">
-                {project_name}
-              </span>
-              <span className="font-mono opacity-50">{organization_name}</span>
-            </div>
-            {project && (
+      <div className="flex flex-col overflow-hidden w-full h-full">
+        <DesktopDragArea className="border-b" />
+        <main className="w-full h-full overflow-y-scroll">
+          <div className="container mx-auto">
+            <header className="py-10 flex justify-between">
               <div>
-                <CreateNewDocumentButton
+                <span className="flex items-center gap-2 text-2xl font-black select-none">
+                  {project_name}
+                </span>
+                <span className="font-mono opacity-50">
+                  {organization_name}
+                </span>
+              </div>
+              {project && (
+                <div>
+                  <CreateNewDocumentButton
+                    project_name={project_name}
+                    project_id={project.id}
+                  />
+                </div>
+              )}
+            </header>
+            {loading ? (
+              <>
+                <div>
+                  <Skeleton className="w-full h-32 rounded" />
+                </div>
+              </>
+            ) : (
+              <div>
+                <section>
+                  <ProjectStats project_ids={[project!.id]} />
+                </section>
+                <section className="w-full flex justify-end gap-2 mt-10">
+                  <Link href="?layout=grid" replace>
+                    <ViewGridIcon />
+                  </Link>
+                  <Link href="?layout=list" replace>
+                    <ViewHorizontalIcon />
+                  </Link>
+                </section>
+                <hr className="mb-10 mt-5 dark:border-neutral-700" />
+                <DocumentsGrid
+                  organization_name={organization_name}
                   project_name={project_name}
-                  project_id={project.id}
+                  documents={documents.filter(
+                    (doc) => doc.project_id === project!.id
+                  )}
+                  layout={layout}
                 />
+                <footer className="mt-10 mb-5">
+                  <PoweredByGridaFooter />
+                </footer>
               </div>
             )}
-          </header>
-          {loading ? (
-            <>
-              <div>
-                <Skeleton className="w-full h-32 rounded" />
-              </div>
-            </>
-          ) : (
-            <div>
-              <section>
-                <ProjectStats project_ids={[project!.id]} />
-              </section>
-              <section className="w-full flex justify-end gap-2 mt-10">
-                <Link href="?layout=grid" replace>
-                  <ViewGridIcon />
-                </Link>
-                <Link href="?layout=list" replace>
-                  <ViewHorizontalIcon />
-                </Link>
-              </section>
-              <hr className="mb-10 mt-5 dark:border-neutral-700" />
-              <DocumentsGrid
-                organization_name={organization_name}
-                project_name={project_name}
-                documents={documents.filter(
-                  (doc) => doc.project_id === project!.id
-                )}
-                layout={layout}
-              />
-              <footer className="mt-10 mb-5">
-                <PoweredByGridaFooter />
-              </footer>
-            </div>
-          )}
-        </div>
-      </main>
+          </div>
+        </main>
+      </div>
     </div>
   );
 }
