+++ editor/app/(workspace)/[org]/page.tsx
@@ -26,6 +26,7 @@ import { WorkspaceSidebar } from "@/scaffolds/workspace/sidebar";
 import { useWorkspace } from "@/scaffolds/workspace";
 import { Skeleton } from "@/components/ui/skeleton";
 import { editorlink } from "@/lib/forms/url";
+import { DesktopDragArea } from "@/components/desktop-drag-area";
 
 export default function DashboardProjectsPage({
   params,
@@ -46,92 +47,95 @@ export default function DashboardProjectsPage({
   return (
     <div className="h-full flex flex-1 w-full">
       <WorkspaceSidebar />
-      <main className="w-full h-full overflow-y-scroll">
-        <div className="container mx-auto">
-          <header className="py-10">
-            <div>
-              <span className="text-2xl font-black">Home</span>
-            </div>
-          </header>
-          {loading ? (
-            <></>
-          ) : (
-            <>
-              <section className="py-10">
-                <ProjectStats project_ids={projects.map((p) => p.id)} />
-              </section>
-            </>
-          )}
-          <section className="w-full flex justify-end gap-2 mt-10">
-            <Link href="?layout=grid" replace>
-              <ViewGridIcon />
-            </Link>
-            <Link href="?layout=list" replace>
-              <ViewHorizontalIcon />
-            </Link>
-          </section>
-          <hr className="mb-10 mt-5" />
-          {loading ? (
-            <ProjectsLoading />
-          ) : (
-            <>
-              {projects.length === 0 && (
-                <Card>
-                  <CardContent className="py-16">
-                    <CardHeader />
-                    <div className="flex flex-col items-center justify-center gap-8">
-                      <div className="flex flex-col gap-2 items-center">
-                        <BoxSelectIcon className="w-12 h-12 text-muted-foreground" />
-                        <h2 className="text-lg text-muted-foreground">
-                          No projects yet
-                        </h2>
+      <div className="flex flex-col overflow-hidden w-full h-full">
+        <DesktopDragArea className="border-b" />
+        <main className="w-full h-full overflow-y-scroll">
+          <div className="container mx-auto">
+            <header className="py-10">
+              <div>
+                <span className="text-2xl font-black">Home</span>
+              </div>
+            </header>
+            {loading ? (
+              <></>
+            ) : (
+              <>
+                <section className="py-10">
+                  <ProjectStats project_ids={projects.map((p) => p.id)} />
+                </section>
+              </>
+            )}
+            <section className="w-full flex justify-end gap-2 mt-10">
+              <Link href="?layout=grid" replace>
+                <ViewGridIcon />
+              </Link>
+              <Link href="?layout=list" replace>
+                <ViewHorizontalIcon />
+              </Link>
+            </section>
+            <hr className="mb-10 mt-5" />
+            {loading ? (
+              <ProjectsLoading />
+            ) : (
+              <>
+                {projects.length === 0 && (
+                  <Card>
+                    <CardContent className="py-16">
+                      <CardHeader />
+                      <div className="flex flex-col items-center justify-center gap-8">
+                        <div className="flex flex-col gap-2 items-center">
+                          <BoxSelectIcon className="w-12 h-12 text-muted-foreground" />
+                          <h2 className="text-lg text-muted-foreground">
+                            No projects yet
+                          </h2>
+                        </div>
+                        <CreateNewProjectDialog org={organization.name}>
+                          <Button variant="outline">
+                            <PlusIcon className="inline w-4 h-4 me-2" />
+                            Create your first project
+                          </Button>
+                        </CreateNewProjectDialog>
                       </div>
-                      <CreateNewProjectDialog org={organization.name}>
-                        <Button variant="outline">
-                          <PlusIcon className="inline w-4 h-4 me-2" />
-                          Create your first project
-                        </Button>
-                      </CreateNewProjectDialog>
-                    </div>
-                    <CardFooter />
-                  </CardContent>
-                </Card>
-              )}
-              {projects.map((p) => {
-                const projectdocuments = documents.filter(
-                  (d) => d.project_id === p.id
-                );
+                      <CardFooter />
+                    </CardContent>
+                  </Card>
+                )}
+                {projects.map((p) => {
+                  const projectdocuments = documents.filter(
+                    (d) => d.project_id === p.id
+                  );
 
-                return (
-                  <div key={p.id} className="mb-40">
-                    <header className="sticky top-0 py-5 mb-10 flex justify-between items-center border-b bg-background z-10">
-                      <Link href={`/${organization.name}/${p.name}`}>
-                        <h2 className="text-2xl font-bold">
-                          {p.name}
-                          <OpenInNewWindowIcon className="inline align-middle ms-2 w-5 h-5" />
-                        </h2>
-                      </Link>
-                      <CreateNewDocumentButton
+                  return (
+                    <div key={p.id} className="mb-40">
+                      <header className="sticky top-0 py-5 mb-10 flex justify-between items-center border-b bg-background z-10">
+                        <Link href={`/${organization.name}/${p.name}`}>
+                          <h2 className="text-2xl font-bold">
+                            {p.name}
+                            <OpenInNewWindowIcon className="inline align-middle ms-2 w-5 h-5" />
+                          </h2>
+                        </Link>
+                        <CreateNewDocumentButton
+                          project_name={p.name}
+                          project_id={p.id}
+                        />
+                      </header>
+                      <DocumentsGrid
+                        organization_name={organization.name}
                         project_name={p.name}
-                        project_id={p.id}
+                        documents={projectdocuments}
+                        layout={layout}
                       />
-                    </header>
-                    <DocumentsGrid
-                      organization_name={organization.name}
-                      project_name={p.name}
-                      documents={projectdocuments}
-                      layout={layout}
-                    />
-                  </div>
-                );
-              })}
-            </>
-          )}
-          <footer className="mt-10 mb-5">
-            <PoweredByGridaFooter />
-          </footer>
-        </div>
-      </main>
+                    </div>
+                  );
+                })}
+              </>
+            )}
+            <footer className="mt-10 mb-5">
+              <PoweredByGridaFooter />
+            </footer>
+          </div>
+        </main>
+      </div>
     </div>
   );
 }
