--- editor/app/desktop.css
@@ -1,3 +1,0 @@
-.desktop-drag-area{
-  app-region: drag;
-}
