+++ editor/components/sidebar/index.tsx
@@ -8,6 +8,7 @@ import Link from "next/link";
 import { CaretDownIcon, CaretRightIcon } from "@radix-ui/react-icons";
 import { FixedSizeGrid, type GridChildComponentProps } from "react-window";
 import { useMeasure } from "@uidotdev/usehooks";
+import { DesktopDragArea } from "../desktop-drag-area";
 
 export function SidebarRoot({
   side = "left",
@@ -30,6 +31,10 @@ export function SidebarRoot({
       {children}
     </aside>
   );
+}
+
+export function SidebarDesktopHeader() {
+  return <DesktopDragArea className="border-b" />;
 }
 
 export function SidebarHeader({
